starting ui lobbies
