+++ tailwind.config.js
@@ -8,6 +8,8 @@ module.exports = {
   theme: {
     extend: {
       width: {
+        "7/10": "70%",
+        "2.7/10": "27%",
         "1/10": "10%",
         "1/11": "9.09090909%",
         "1/13": "7.69230769%",
